Enum osqp::Status

pub enum Status<'a> {
    Solved(Solution<'a>),
    SolvedInaccurate(Solution<'a>),
    MaxIterationsReached(Solution<'a>),
    TimeLimitReached(Solution<'a>),
    PrimalInfeasible(PrimalInfeasibilityCertificate<'a>),
    PrimalInfeasibleInaccurate(PrimalInfeasibilityCertificate<'a>),
    DualInfeasible(DualInfeasibilityCertificate<'a>),
    DualInfeasibleInaccurate(DualInfeasibilityCertificate<'a>),
    NonConvex(Failure<'a>),
    // some variants omitted
}

The result of solving a problem.

Variants

Solved

Tuple Fields of Solved

0: Solution<'a>

SolvedInaccurate

Tuple Fields of SolvedInaccurate

0: Solution<'a>

MaxIterationsReached

Tuple Fields of MaxIterationsReached

0: Solution<'a>

TimeLimitReached

Tuple Fields of TimeLimitReached

0: Solution<'a>

PrimalInfeasible

Tuple Fields of PrimalInfeasible

0: PrimalInfeasibilityCertificate<'a>

PrimalInfeasibleInaccurate

Tuple Fields of PrimalInfeasibleInaccurate

0: PrimalInfeasibilityCertificate<'a>

DualInfeasible

Tuple Fields of DualInfeasible

0: DualInfeasibilityCertificate<'a>

DualInfeasibleInaccurate

Tuple Fields of DualInfeasibleInaccurate

0: DualInfeasibilityCertificate<'a>

NonConvex

Tuple Fields of NonConvex

0: Failure<'a>

Implementations

impl<'a> Status<'a>

pub fn x(&self) -> Option<&'a [f64]>

Returns the primal variables at the solution if the problem is Solved.

pub fn solution(&self) -> Option<Solution<'a>>

Returns the solution if the problem is Solved.

pub fn iter(&self) -> u32

Returns the number of iterations taken by the solver.

pub fn setup_time(&self) -> Duration

Returns the time taken for the setup phase.

pub fn solve_time(&self) -> Duration

Returns the time taken for the solve phase.

pub fn polish_time(&self) -> Duration

Returns the time taken for the polish phase.

pub fn run_time(&self) -> Duration

Returns the total time taken by the solver.

This includes the time taken for the setup phase on the first solve.

pub fn rho_updates(&self) -> u32

Returns the number of rho updates.

pub fn rho_estimate(&self) -> f64

Returns the current best estimate of rho.