Struct osqp::Settings

pub struct Settings { /* fields omitted */ }

The settings used when initialising a solver.

Methods

impl Settings

pub fn rho(self, value: f64) -> Settings

Sets the ADMM step rho.

pub fn sigma(self, value: f64) -> Settings

Sets the ADMM step sigma.

pub fn scaling(self, value: Option<u32>) -> Settings

Sets the number of heuristic data scaling iterations.

If None scaling is disabled.

Panics on 32-bit platforms if the value is above i32::max_value().

pub fn adaptive_rho(self, value: bool) -> Settings

Enables choosing rho adaptively.

pub fn adaptive_rho_interval(self, value: Option<u32>) -> Settings

Sets the number of iterations between rho adaptations.

If None it is automatic.

Panics on 32-bit platforms if the value is above i32::max_value().

pub fn adaptive_rho_tolerance(self, value: f64) -> Settings

Sets the tolerance for adapting rho.

The new rho has to be x times larger or 1/x times smaller than the current rho to trigger a new factorization.

pub fn adaptive_rho_fraction(self, value: f64) -> Settings

Set the interval for adapting rho as a fraction of the setup time.

pub fn max_iter(self, value: u32) -> Settings

Sets the maximum number of ADMM iterations.

Panics on 32-bit platforms if the value is above i32::max_value().

pub fn eps_abs(self, value: f64) -> Settings

Sets the absolute convergence tolerance.

pub fn eps_rel(self, value: f64) -> Settings

Sets the relative convergence tolerance.

pub fn eps_prim_inf(self, value: f64) -> Settings

Sets the primal infeasibility tolerance.

pub fn eps_dual_inf(self, value: f64) -> Settings

Sets the dual infeasibility tolerance.

pub fn alpha(self, value: f64) -> Settings

Sets the linear solver relaxation parameter.

pub fn linsys_solver(self, value: LinsysSolver) -> Settings

Sets the linear system solver to use.

pub fn delta(self, value: f64) -> Settings

Sets the polishing regularization parameter.

pub fn polish(self, value: bool) -> Settings

Enables polishing the ADMM solution.

pub fn polish_refine_iter(self, value: u32) -> Settings

Sets the number of iterative refinement steps to use when polishing.

Panics on 32-bit platforms if the value is above i32::max_value().

pub fn verbose(self, value: bool) -> Settings

Enables writing progress to stdout.

pub fn scaled_termination(self, value: bool) -> Settings

Enables scaled termination criteria.

pub fn check_termination(self, value: Option<u32>) -> Settings

Sets the number of ADMM iterations between termination checks.

If None termination checking is disabled.

Panics on 32-bit platforms if the value is above i32::max_value().

pub fn warm_start(self, value: bool) -> Settings

Enables warm starting the primal and dual variables from the previous solution.

pub fn time_limit(self, value: Option<Duration>) -> Settings

Sets the solve time limit.